Prepare CCseqBasic.sh for bigger datasets & mapq 20
CCseqBasic.sh was calling fastqc with only one thread and thus only 250mb of memory. For the new TriC data set this lead to a memory overflow. Increasing the amount of threads to two resolved this issue. The only consequence is that it will now need min 500mb of memory instead of 250mb. This should not be an issue as this pipeline is never run with such little memory.

Also, the new standard for mapq is 20. The "scratches" still don't show and some more data points can be retained.