new normalization
In general all matrices are now normed to 300,000 total interactions. When the Sum of all interactions does not fit 300k, a factor is calculated and every datapoint is multiplied by that factor to get to 300k.
It is also possible to specify a normchrom, normstart and normstop. If specified, this region will be used for normalization instead of the region that is being plotted. This is useful to still have the same normalization in the zoom in plots as in the whole matrices.